Move GLTFExporter.js and phoenixExport.js to lib/
